refactor(SimpleFeederSystem): Move to more organized directory
Due to the potential amount of concrete variants, each blueprint will have their own directory with their concrete makers/factories
