+++ fang-mcb-project/src/control/turret/feeder/simple_feeder_system/m2006_simple_feeder_maker.hpp
@@ -1,6 +1,6 @@
-#ifndef FANG_ROBOTICS_MCB_CONTROL_TURRET_SYSTEM_MAKER_M2006_SIMPLE_FEEDER_MAKER_HPP
-#define FANG_ROBOTICS_MCB_CONTROL_TURRET_SYSTEM_MAKER_M2006_SIMPLE_FEEDER_MAKER_HPP
-#include "control/turret/system/blueprint/simple_feeder_system.hpp"
+#ifndef FANG_ROBOTICS_MCB_CONTROL_TURRET_FEEDER_SIMPLE_FEEDER_M2006_SIMPLE_FEEDER_MAKER_HPP
+#define FANG_ROBOTICS_MCB_CONTROL_TURRET_FEEDER_SIMPLE_FEEDER_M2006_SIMPLE_FEEDER_MAKER_HPP
+#include "simple_feeder_system.hpp"
 #include "drivers.hpp"
 #include "trap/motor/dji_m2006.hpp"
